Move unescape code to macro-support
